backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.10.0
sse-starlette>=1.8.2
